Stage one of the BSR: Initial setup
After much fiddling around, I think I've hit on a method to have the
source of the package split across multiple source files that is not
overly ugly and is only slightly hack-ish.

The idea is that all source files save one will be outside of the root
of the package directory, with that sole plugin file being responsible
for loading the other plugin files and exposing them to Sublime so
that the commands are available.

The sole plugin file should also ensure that, when it is loaded by
Sublime, any of the other package files that might have previously
been loaded are reloaded so that any new changes are picked up.

To this end, there is a function named _reload() in the top level
plugin file (imported from there where needed) which takes the name of
a submodule in the OverrideAudit name space and an optional list of
submodules of that module, and reloads any that may already have been
loaded.

In addition to this, each folder that represents a submodule of some
sort has an __init__.py source file in it that uses the _reload()
method to ensure that all of its contents have been reloaded, then
imports everything and exports that back to whoever imported it.

The top level plugin can then trigger a simple reload of the top
modules (which will recursively reload their children), and then
import both of them to bring all of the command classes in so that
Sublime can see them.

One last modification to this is that sublime will not, in some cases,
invoke plugin_loaded() and plugin_unloaded() in submodules, depending
on the situation. In order to ensure that things always work as
expected, the top level plugin proxies those events, which are handled
by functions with different names to ensure that we have full control
over when they are called.

This first step in the BSR, besides having this extensive log message,
also sets up the first parts of the re-factor by pulling a single
command out as a template for how the rest of the commands will
follow. The extracted command is responsible for changing the mark on
overrides during freshening. It was chosen because it was the least
invasive command to extract.
