+++ lib/threads.py
@@ -1,6 +1,7 @@
 import sublime
 import threading
 
+
 ###----------------------------------------------------------------------------
 
 
@@ -17,6 +18,7 @@ _spinner_key.lock = threading.Lock()
 
 
 ###----------------------------------------------------------------------------
+
 
 class Spinner():
     """
