+++ src/commands/__init__.py
@@ -1,0 +1,9 @@
+from ...bootstrap import reload
+
+reload("src.commands", ["modify_mark"])
+
+from .modify_mark import OverrideAuditModifyMarkCommand
+
+__all__ = [
+    "OverrideAuditModifyMarkCommand"
+]
